+++ test/e2e/specs/publish-button.test.js
@@ -22,22 +22,23 @@ describe( 'PostPublishButton', () => {
 	} );
 
 	it( 'should be disabled when post is not saveable', async () => {
-		const publishButton = await page.$( '.editor-post-publish-button:disabled' );
-
+		const publishButton = await page.$( '.editor-post-publish-button[aria-disabled="true"]' );
 		expect( publishButton ).not.toBeNull();
 	} );
 
 	it( 'should be disabled when post is being saved', async () => {
 		await page.type( '.editor-post-title__input', 'E2E Test Post' ); // Make it saveable
-		expect( await page.$( '.editor-post-publish-button:disabled' ) ).toBeNull();
+		expect( await page.$( '.editor-post-publish-button[aria-disabled="true"]' ) ).toBeNull();
+
 		await page.click( '.editor-post-save-draft' );
-		expect( await page.$( '.editor-post-publish-button:disabled' ) ).not.toBeNull();
+		expect( await page.$( '.editor-post-publish-button[aria-disabled="true"]' ) ).not.toBeNull();
 	} );
 
 	it( 'should be disabled when metabox is being saved', async () => {
 		await page.type( '.editor-post-title__input', 'E2E Test Post' ); // Make it saveable
-		expect( await page.$( '.editor-post-publish-button:disabled' ) ).toBeNull();
+		expect( await page.$( '.editor-post-publish-button[aria-disabled="true"]' ) ).toBeNull();
+
 		await page.evaluate( () => window.wp.data.dispatch( 'core/edit-post' ).requestMetaBoxUpdates() );
-		expect( await page.$( '.editor-post-publish-button:disabled' ) ).not.toBeNull();
+		expect( await page.$( '.editor-post-publish-button[aria-disabled="true"]' ) ).not.toBeNull();
 	} );
 } );